restructure files and folders for my policies module
